requirements.txt
streamlit==1.42.0
finnhub-python==2.4.22